requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
openpyxl
python-pptx
pymupdf
pdf2docx
python-docx